feat: add GuestAnswersModal component and integrate QR code generation
- Introduced GuestAnswersModal for displaying guest form answers.
- Updated GuestListSection to utilize GuestAnswersModal.
- Implemented QR code generation for guest tickets using the qrcode library.
- Refactored guest actions into separate guest-actions.ts file for better organization.
- Added necessary type definitions for qrcode library in package.json.
