+++ frontend/next.config.ts
@@ -8,6 +8,16 @@ const nextConfig: NextConfig = {
   turbopack: {
     root: dirname,
   },
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'fspmilcylelzpzumxvis.supabase.co',
+        port: '',
+        pathname: '/storage/v1/object/public/**',
+      },
+    ],
+  },
 };
 
 export default nextConfig;
