+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@supabase/ssr": "^0.8.0",
     "@tailwindcss/typography": "^0.5.19",
     "@types/leaflet": "^1.9.21",
+    "@types/qrcode": "^1.5.6",
     "clsx": "^2.1.1",
     "dompurify": "^3.3.1",
     "driver.js": "^1.4.0",
@@ -21,6 +22,7 @@
     "nodemailer": "^7.0.12",
     "nunjucks": "^3.2.4",
     "papaparse": "^5.5.3",
+    "qrcode": "^1.5.4",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "recharts": "^3.7.0",
